+++ src/main/java/roomescape/auth/service/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.auth.service.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
