+++ src/main/java/roomescape/auth/service/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto.response;
+package roomescape.auth.service.dto.response;
 
 public record LoginResponse(
         String tokenValue
