+++ src/main/java/roomescape/waiting/controller/WaitingController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.annotation.RequiredAdmin;
-import roomescape.auth.dto.LoginMember;
+import roomescape.auth.service.dto.LoginMember;
 import roomescape.waiting.service.WaitingService;
 import roomescape.waiting.service.dto.request.CreateWaitingRequest;
 import roomescape.waiting.service.dto.response.CreateWaitingResponse;
