+++ src/test/java/roomescape/waiting/service/WaitingServiceTest.java
@@ -7,8 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
-import roomescape.auth.dto.LoginMember;
-import roomescape.common.exception.EntityNotFoundException;
+import roomescape.auth.service.dto.LoginMember;
 import roomescape.common.exception.ForbiddenException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
